+++ sentry-samples/sentry-samples-spring-boot/src/main/java/io/sentry/samples/spring/boot/PersonService.java
@@ -3,6 +3,7 @@ package io.sentry.samples.spring.boot;
 import io.sentry.spring.tracing.SentrySpan;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Service;
 
 /**
@@ -14,12 +15,19 @@ import org.springframework.stereotype.Service;
 public class PersonService {
   private static final Logger LOGGER = LoggerFactory.getLogger(PersonService.class);
 
+  private final JdbcTemplate jdbcTemplate;
+
+  public PersonService(JdbcTemplate jdbcTemplate) {
+    this.jdbcTemplate = jdbcTemplate;
+  }
+
+  @SentrySpan
   Person create(Person person) {
     LOGGER.warn("Creating person: {}", person);
-    try {
-      Thread.sleep(100);
-    } catch (InterruptedException e) {
-    }
+    jdbcTemplate.update(
+        "insert into person (firstName, lastName) values (?, ?)",
+        person.getFirstName(),
+        person.getLastName());
     return person;
   }
 }